+++ src/com/donny/dendrofinance/data/backingtable/AccountBTC.java
@@ -53,18 +53,16 @@ public class AccountBTC extends BackingTableCore<Account> {
         ArrayList<String[]> out = new ArrayList<>();
         for (Account a : TABLE) {
             String check = search;
-            boolean flag = false, rFlag = false, allow = true;
+            boolean flag = a.inUse(), allow = true;
             if (check.contains("$U")) {
-                flag = true;
                 check = check.replace("$U", "").trim();
-                if (!a.inUse()) {
+                if (!flag) {
                     allow = false;
                 }
             }
             if (check.contains("$u")) {
-                rFlag = true;
                 check = check.replace("$u", "").trim();
-                if (a.inUse()) {
+                if (flag) {
                     allow = false;
                 }
             }
@@ -81,7 +79,7 @@ public class AccountBTC extends BackingTableCore<Account> {
                 out.add(new String[]{
                         "" + a.getAid(), a.getName(), a.getCurrency().getName(), a.getAccountType().NAME,
                         a.getBroadAccountType().toString(), a.getBudgetType(), !a.EXPORT ? "X" : "",
-                        (!rFlag && (flag || a.inUse())) ? "X" : ""
+                        flag ? "X" : ""
                 });
             }
         }