+++ src/com/donny/dendrofinance/data/backingtable/CurrencyBTC.java
@@ -55,32 +55,29 @@ public class CurrencyBTC extends BackingTableCore<LCurrency> {
         ArrayList<String[]> out = new ArrayList<>();
         for (LCurrency cur : TABLE) {
             String check = search;
-            boolean flagU = false, rFlagU = false, flagA = false, rFlagA = false, allow = true;
+            boolean flagU = cur.inUse(), allow = true;
+            boolean flagA = flagU || cur.inAccount();
             if (check.contains("$U")) {
-                flagU = true;
                 check = check.replace("$U", "").trim();
-                if (!cur.inUse()) {
+                if (!flagU) {
                     allow = false;
                 }
             }
             if (check.contains("$u")) {
-                rFlagU = true;
                 check = check.replace("$u", "").trim();
-                if (cur.inUse()) {
+                if (flagU) {
                     allow = false;
                 }
             }
             if (check.contains("$A")) {
-                flagA = true;
                 check = check.replace("$A", "").trim();
-                if (!cur.inAccount()) {
+                if (!flagA) {
                     allow = false;
                 }
             }
             if (check.contains("$a")) {
-                rFlagA = true;
                 check = check.replace("$a", "").trim();
-                if (cur.inAccount()) {
+                if (flagA) {
                     allow = false;
                 }
             }
@@ -106,8 +103,8 @@ public class CurrencyBTC extends BackingTableCore<LCurrency> {
             if (allow) {
                 out.add(new String[]{
                         name.toString(), cur.getTicker(), cur.encode(BigDecimal.ZERO), type.toString(),
-                        (!rFlagA && (flagA || cur.inAccount())) ? "X" : "",
-                        (!rFlagU && (flagU || cur.inUse())) ? "X" : ""
+                        flagA ? "X" : "",
+                        flagU ? "X" : ""
                 });
             }
         }